requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
motor==3.7.1
multidict==6.7.0
propcache==0.4.1
pymongo==4.15.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8.1
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
